Missing phenotypes are now allowed
Modified the kernel so that at each individual we check that the geno
isn't missing AND the pheno isn't missing before incrementing group
means etc.

Implemented a small routine in the simqtl function in readplink.c that
asks the user how many missing values to include in the phenotype array,
and then randomly samples based on seed set by srand(time(0)).

The df2 column corresponds to what we expect with missing values,
indicating that the kernel is functioning correctly.