binary plink still not working
The 16th genotype in each int seems to be complement to the correct
value, manually overriding this, but I don't know why it is happening.

The 'pack' function takes 012 format geno array in individual-major
format. However, the conversion from binary plink to 012 is in SNP major
order, so this is problematic.